Fix typescript types for useExtraDeps
The types were basically allowing `any` to be used throughout due to the
type alias for `CallbackFn`

The orginal flow formulation used opaque types for `CallbackFn` to avoid
this issue. TS does not have opaque types so a record was used instead.

This change will require some changes in app code, namely that type
inference will infer an `unknown` type for the dependencies. This is
annyoing, but it is typesafe.

We will need to annotate the generic param when calling any of the
`useSafe*` functions, see the tests for how it was done.